+++ packages/forms/src/locales/en.ts
@@ -6,6 +6,10 @@ export default {
     date: 'The field must be a valid date',
     decrease: 'Decrease',
     email: 'The field must be a valid email address.',
+    hasMixed:
+      'The field must contain at least one uppercase and one lowercase letter.',
+    hasNumber: 'The field must contain at least one number.',
+    hasSymbol: 'The field must contain at least one symbol.',
     increase: 'Increase',
     integer: 'The field must be an integer.',
     maxChars: 'The field must not be greater than {{max}} characters.',