+++ packages/forms/src/locales/it.ts
@@ -6,6 +6,10 @@ export default {
     date: 'Il campo deve essere una data valida',
     decrease: 'Diminuisci',
     email: 'Il campo deve essere un indirizzo email valido.',
+    hasMixed:
+      'Il campo deve contenere almeno un carattere maiuscolo ed un carattere minuscolo.',
+    hasNumber: 'Il campo deve contenere almeno un numero.',
+    hasSymbol: 'Il campo deve contenere almeno un simbolo.',
     increase: 'Aumenta',
     integer: 'Il campo deve essere un numero intero.',
     maxChars: 'Il campo non può contenere più di {{max}} caratteri.',