+++ packages/tsv-advanced/src/rules/email.ts
@@ -1,4 +1,5 @@
-import { Message, emailRegex } from '@gmcode/tsv-core'
+import { Message } from '@gmcode/tsv-core'
+import { isEmail } from '@gmcode/tsv-core/regex'
 import { AdvancedRule } from '@/advanced-rule'
 import type { AdvancedRuleOptions, ParseEmpty } from '@/types'
 
@@ -85,7 +86,7 @@ export class EmailRule<
     }
 
     // Email regex
-    if (!emailRegex.test(value)) {
+    if (!isEmail.test(value)) {
       return new Message(this.messages.email)
     }