+++ vite.config.ts
@@ -8,7 +8,8 @@ export default defineConfig({
       'pnpm-lock.yaml',
       '**/dist',
       'internal/mock-inertia/core.js',
-      '**/mockServiceWorker.js'
+      '**/mockServiceWorker.js',
+      '.changeset/*.md'
     ],
     overrides: [
       {